Refactoring data for windowing
