+++ lib/Windowing/stories/WindowingStory.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, number } from '@storybook/addon-knobs';
-import uuid from 'uuid/v4';
 import {
   WindowingProvider,
   InViewList,
@@ -11,6 +10,7 @@ import {
 } from '../../index';
 import { WindowingFolderRow } from './WindowingFolderRow';
 import { WindowingNestedContainer } from './WindowingNestedContainer';
+import { createData } from './windowingData';
 
 const stories = storiesOf('Components/Windowing', module);
 stories.addDecorator(withKnobs);
@@ -42,36 +42,10 @@ stories
     );
   })
   .add('Nested V2', () => {
-    const parentIds = [...new Array(5)].map(() => uuid());
-    const byParentIds = parentIds.reduce(
-      (acc, value) => ({
-        ...acc,
-        [value]: [...new Array(5)].map(() => uuid())
-      }),
-      {}
-    );
+    const parents = number('Number of parents', 5);
+    const children = number('Number of children each parent has ', 5);
 
-    const allIds = parentIds.reduce(
-      (acc, value) => [...acc, value, ...byParentIds[value]],
-      []
-    );
-
-    let depth = 0;
-    const byId = allIds.reduce((acc, value, index) => {
-      const lastId = allIds[index - 1];
-      if (parentIds.indexOf(lastId) > -1) {
-        depth += 1;
-      }
-
-      return {
-        ...acc,
-        [value]: {
-          id: value,
-          name: parentIds.indexOf(value) > -1 ? 'Folder' : 'Item',
-          depth
-        }
-      };
-    }, {});
+    const { allIds, parentIds, byId } = createData(parents, children);
 
     return (
       <WindowingNestedContainer allIds={allIds}>
